+++ src/g_qt.cpp
@@ -191,7 +191,7 @@ public:
     ~GFramePeer();
     
     int ConSetTitle(char *Title, char *STitle);
-    int ConGetTitle(char *Title, int MaxLen, char *STitle, int SMaxLen);
+    int ConGetTitle(char *Title, size_t MaxLen, char *STitle, size_t SMaxLen);
     
     int ConSetSize(int X, int Y);
     int ConQuerySize(int *X, int *Y);
@@ -1479,7 +1479,7 @@ int GFramePeer::ConSetTitle(char *Title, char *STitle) {
     return 1;
 }
 
-int GFramePeer::ConGetTitle(char *Title, int MaxLen, char *STitle, int SMaxLen) {
+int GFramePeer::ConGetTitle(char *Title, size_t MaxLen, char *STitle, size_t SMaxLen) {
     strncpy(Title, qFrame->caption(), MaxLen);
     strncpy(STitle, qFrame->iconText(), SMaxLen);
     return 1;
@@ -1559,7 +1559,7 @@ int GFrame::ConSetTitle(char *Title, char *STitle) {
     return Peer->ConSetTitle(Title, STitle);
 }
 
-int GFrame::ConGetTitle(char *Title, int MaxLen, char *STitle, int SMaxLen) {
+int GFrame::ConGetTitle(char *Title, size_t MaxLen, char *STitle, size_t SMaxLen) {
     return Peer->ConGetTitle(Title, MaxLen, STitle, SMaxLen);
 }
 
@@ -2052,7 +2052,7 @@ int GUI::SetPipeView(int id, EModel *notify) {
     return 0;
 }
 
-ssize_t GUI::ReadPipe(int id, void *buffer, int len) {
+ssize_t GUI::ReadPipe(int id, void *buffer, size_t len) {
     ssize_t rc;
     
     if (id < 0 || id > MAX_PIPES)
